Updated project cluster mapping
